+++ src/scmdata/__init__.py
@@ -7,4 +7,4 @@ __version__ = get_versions()["version"]
 del get_versions
 
 from scmdata.dataframe import ScmDataFrame, df_append  # noqa: F401, E402
-from scmdata.run import ScmRun  # noqa: F401, E402
+from scmdata.run import ScmRun, run_append  # noqa: F401, E402